refactor: move toad.py to cli package
